Added the half finished GUI prototype to the branch
The GUI has been re-written to be compatible with the new HybRIDS
object system and is halfway to being complete - users can make a
HybRIDS object (called a HybRIDS session in the GUI) they can also
generate triplets to be analyzed, and analyze the seuqence similarity,
find potential blocks. Next up dating the potential blocks, the
plotting pane, and export of data.