Changes to sequence input
Changed how sequences are input so now the format of an input file is
infered from the filename extension i.e. after the "." alternativeley,
the user can force an extention or give the name of a compatible object
in place of a filename and if possible, HybRIDS will convert interpret
it and read it into a HybRIDSdna object.

This change is in preparation for more future support for more file
formats - we start with the FASTA file format support, and also support
for the DNAbin object.