Altered duplication detection process
Now the duplication detection process uses the ape function dist.dna to
very rapidly compute a matrix of pairwise raw distances, we then simply
find the ones that are equal to 0 - indicating duplicate pairs and we
remove one of the pairs from the set of sequences. This will hopefully
be faster than the previous method.